requirements.txt
python-telegram-bot>=21.0
requests>=2.28
